refactor(httpclient): split 1,652-line file into 9 focused modules
Split src/widget/httpclient.rs (1,652 lines) into 9 submodules:
- types.rs: HttpMethod, RequestState, ContentType, ResponseView, HttpColors
- request.rs: HttpRequest struct with builder pattern
- response.rs: HttpResponse with JSON formatting
- builder.rs: RequestBuilder for fluent API and base64_encode
- backend.rs: HttpBackend trait and MockHttpBackend
- client.rs: HttpClient main widget struct
- render.rs: View trait implementation
- helpers.rs: Convenience functions (get, post, put, delete, patch)
- tests.rs: All 44 tests in separate module

Reduced main file from 1,652 to 53 lines (97% reduction).

Changes:
- Separated concerns into focused modules following SRP
- Made format_json() public for test access
- Fixed explicit counter loop in render.rs
- Added doc-hidden to base64_encode (internal helper)
- All 44 tests pass, clippy clean
