+++ src/widget/httpclient/request.rs
@@ -1,0 +1,68 @@
+//! HTTP request types
+
+use std::collections::HashMap;
+
+use super::types::HttpMethod;
+
+/// HTTP request builder
+#[derive(Clone, Debug, Default)]
+pub struct HttpRequest {
+    /// HTTP method
+    pub method: HttpMethod,
+    /// Request URL
+    pub url: String,
+    /// Request headers
+    pub headers: HashMap<String, String>,
+    /// Request body
+    pub body: String,
+    /// Query parameters
+    pub params: HashMap<String, String>,
+}
+
+impl HttpRequest {
+    /// Create a new request
+    pub fn new(url: impl Into<String>) -> Self {
+        Self {
+            url: url.into(),
+            ..Default::default()
+        }
+    }
+
+    /// Set method
+    pub fn method(mut self, method: HttpMethod) -> Self {
+        self.method = method;
+        self
+    }
+
+    /// Add header
+    pub fn header(mut self, key: impl Into<String>, value: impl Into<String>) -> Self {
+        self.headers.insert(key.into(), value.into());
+        self
+    }
+
+    /// Set body
+    pub fn body(mut self, body: impl Into<String>) -> Self {
+        self.body = body.into();
+        self
+    }
+
+    /// Add query parameter
+    pub fn param(mut self, key: impl Into<String>, value: impl Into<String>) -> Self {
+        self.params.insert(key.into(), value.into());
+        self
+    }
+
+    /// Build URL with query parameters
+    pub fn full_url(&self) -> String {
+        if self.params.is_empty() {
+            self.url.clone()
+        } else {
+            let params: Vec<String> = self
+                .params
+                .iter()
+                .map(|(k, v)| format!("{}={}", k, v))
+                .collect();
+            format!("{}?{}", self.url, params.join("&"))
+        }
+    }
+}
